refactor(client): extract small message_item renderers to message/ files
